+++ .gitignore
@@ -1,0 +1,23 @@
+# Dependency directories
+/node_modules
+
+# Environment configuration files
+.env
+
+# Log files
+*.log
+
+# Built output
+/build
+
+# IDE and editor files
+.vscode/
+.idea/
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw*
+
+# Mac specific files
+.DS_Store
