+++ public/index.html
@@ -19,6 +19,8 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.9.359/pdf.min.js"></script>
     <script src="https://html2canvas.hertzen.com/dist/html2canvas.min.js"></script>
     <script src="/meeting.js" type="module"></script>
+    <script src="whiteboard.js" type="module"></script>
+    <script src="screenshare.js" type="module"></script>
 
   </head>
   <body>
